requirements.txt
streamlit
tensorflow
pillow
numpy
matplotlib
pandas
